Uint16 voxel storage: lift the 256-block-id ceiling
Widen per-chunk block storage from Uint8Array to Uint16Array (MAX_BLOCKS
= 1024) so the full catalogue — concrete, terracotta, glazed, redstone,
rails, etc. — fits with headroom. Touches chunk gen, mesher cache,
lighting opacity table, and the BLOCKS registry cap.

Save format bumped to v2: chunks now use a 16-bit RLE codec (count u16 +
value u16, tagged w:16). Legacy byte-RLE records still load and widen to
Uint16, so pre-existing worlds are preserved.

+++ src/render/mesher.js
@@ -27,7 +27,7 @@ const DIR_INFO = [
 const AXIS_VEC = [[1, 0, 0], [0, 1, 0], [0, 0, 1]];
 
 // Scratch buffers, reused between calls (single-threaded meshing).
-const blocksCache = new Uint8Array(P * P * P);
+const blocksCache = new Uint16Array(P * P * P);
 const lightCache = new Uint8Array(P * P * P);
 const mask = new Int32Array(256);
 const maskAO = new Uint8Array(256);
